CHALLENGE: Paging COMPLETE
teh challenge was to detect when the user scrolls to the bottom of the results within the recyclerview, and then retrieve the next page of results from the flickr api

this is very simple. recyclerview can have an OnScrollListener to detect when the end is reached. we merely increment an int representing the current page, then modify FlickrFetchr to take in an int for which page to fetch. the flickr api already has a method for retrieving distinct pages from the getRecent service